Introduced MVC-ish controllers, added abstractions around syncing, and created a skeleton for MVC-ish models.
Originally, we wanted to be able to implement the following: when a new observation is added, the patient chart should be updated. Upon investigation, this proved to be difficult:

 * PatientChartActivity handles the activity result from ODK and posts it to the server.
 * PatientChartFragment displays the observation grid.

We needed PatientChartFragment to wait until the posted succeeds, meaning we need some kind of shared state between the activity and fragment. We could have done this with just an event, but where's the fun in that?