Remove unused code from PatientListFragment and fix minor bug.
ListDialogFragment is never used because the button to open it is
marked as "gone" in the XML. Removed this deprecated class.

mTwoPane is never used -- seems to date from when we were supporting
multiple devices. From what I understand, our plan is to only support
one device -- the Sony tablet we've been using for development --
so we don't need to support multiple screen sizes here.

Also fixed a minor threading issue I noticed -- synchronization on
onEvent in PatientListActivity does more or less nothing; we should
handle events on the main thread instead.