OFK form cancel/done buttons in same style as PatientCreationActivity.
Remove the buttons from the action bar, and add red/green buttons to the
layout in the same style as PatientCreationActivity.

Move some related resources from app to odkcollect so that they can be
shared between the two.

Also simplify the form entry activity layout slightly, to reduce the number
of nested layouts.