Fix patient chart controller unit test.
This was broken by the handler hack in PatientChartController.
Expanded the TODO next to the hack to explain the problem
better, and updated our unit tests to handle the hacky behaviour.

Rather than setting a 100ms delay when posting a runnable to the
handler, it should be fine to post without a delay. Manually tested
that this behaves correctly and modified the code to do this.