Add FastDataGridView to PatientChartActivity
New data grid is behind a property flag for now, but will be enabled by default
shortly.

New implementation is based on 3 RecyclerViews and only one Scrollview.

GridLayoutManager lays out its children in a grid, but only support scrolling in
one of the two dimensions. We put it in a ScrollView to scroll into the second
dimension. This means that recycling will only happen in the horizontal (time)
direction and we still have to inflate each column to its full height.

However:
 - This fixes the main scalability problem of a long patient history.
 - The number of observations is bounded and will rather go a bit down, so the
   height of a column is not too large.
 - This improves inflation time for a larger chart from ~2.5 seconds to 700 ms.
 - It fixes the lagging-behind of the row headers when scrolling the cells.