Use ViewStubs for TextViews in ChartGridView
Most of the time, we don't actually display any text in a cell. So we start out
with using a ViewStub instead and only inflate the TextView once we actually
need it.

This makes scrolling much more fluent and brings the Activity launch time (with
a local change [not commited] to do the inflations directly in onCreate) under
500 ms.