Cleanup: Remove the use of a common <included /> xml file for the tent view.
A style is a better way to share layout across the two places where this include
was used.

Also, delete a redundant FrameLayout and fix up a few lint errors.

SIDE BENEFIT: We can specify layout parameters separately for the two GridViews
now. This change causes the GridView to scroll in both the TentSelectionFragment
and the AssignLocationDialog if there's too many tents. Note that the
TentSelectionFragment as a whole doesn't scroll, just the GridView containing
the tents.