Better type discipline: Obs vs ObsPoint vs ObsValue.
This introduces the concept of a "point" as separate from a "value".
(The name "point" is suggestive of a point on a time series graph.)

ObsValue represents just an observed value (number, text, etc.)
ObsPoint is a new class that represents an ObsValue with a specific time.
Obs identifies the concept being observed as well as the time and value.
(The last step, not yet done, is to rewrite Obs to contain ObsPoint.)

ObsFormat is more consistent in that it always formats ObsValues;
the rendering code in chart.html is also cleaner as a result.

ObsFormat and PebbleExtension are more defensive, avoiding throwing exceptions.