Refactor sync phases into self-contained SyncControllers
Before, there was a huge switch statement in SyncAdapter, which farmed out to individual methods,
which then almost always made a call to a static method on DbSyncUtil. This change pulls the logic
from the different "helper" methods into single, cohesive classes per sync phase.

Note that this commit is almost entirely moving code around, except for in
SyncAdapter#onPerformSync.