Improve reliability of PatientChartActivityTest.
Improve reliability of PatientChartActivityTest by waiting for views to be shown before attempting
to click on them.

I suspect the root cause might be something to do with the way our Activity is structured, or
general flakiness in Espresso, but this resulted in a vast improvement on my setup, pointed at the
dev server.

Note also that this doesn't fix the tests - there are multiple issues that should be addressed in
future commits - but this allows the tests to actually get to the assertions instead of failing in
the setup code.