Make PatientChartActivityTest slightly less flaky.
Previously, testEncounter_latestEncounterIsAlwaysShown would take 3 minutes to run because there was
a 30 second wait and 6 iterations. A 5 second wait is enough for it to work, and 3 iterations is
enough to actually test the same behavior.

Fix an error in testCombinesNonOverlappingObservationsForSameEncounter such that it works with the
test profile and with the dev server.