Generalize sync token storage.
Previously, a server-supplied timestamp was stored for each incremental observation sync. This
timestamp was stored in a field in the "misc" SQLite table.

There are two problems with this:
- Our design (http://j.mp/buendia-sync) advocates the client treating the server-supplied token as
  an opaque sync token instead of as a timestamp
- Soon we'll be storing sync tokens for patients as well.

This change refactors the incremental observation timestamp storage into a more general "sync token"
storage, which can be used my multiple sync phases.