Fix GsonRequest's use of generics.
Previously, GsonRequest didn't support parameterized types. i.e.
- `GsonRequest<MyJsonResponse>` worked, but
- `GsonRequest<MyJsonResponse<OfX>>` didn't.

As a workaround, there was an `isArray` parameter you could pass to the constructor that, if set to true, would cause the type parameter to be interpreted as an array instead of as the base response type. Gross.

This change fixes that problem by moving to using `java.lang.reflect.Type` for the constructor parameter and recommending Gson's `TypeToken` paradigm for this use case. Note that passing `MyJsonResponse.class` to the `type` argument still works.

This change does sacrifice some type-safety, however - `Class<T>` implements `Type`, and so it's entirely possible for code that uses this class to pass the wrong base type for the generic. It would be possible to make this harder by having an overload that accepts `Class<T>`, but unless that parameter was in a different position to the `Type` parameter, this wouldn't achieve much - the `Class<T>` overload would be used _unless_ the wrong class was specified, in which case the wrong class parameter would be interpreted as as `Type`. I didn't think that the extra complication was worth creating an extra overload for, but I'm definitely open to ideas here :)