Improve the way sync progress is reported.
Previously, sync progress was reported by broadcasting Intents with an 'increment' indicating how
much more progress had been made since the last broadcast. Clients were expected to track all
broadcasts, and accumulate the 'increment' amounts locally. This was the paradigm used by the
`SyncAdapter`, which broadcasts Intents to `SyncManager$SyncStatusBroadcastReceiver`, and then
by the `SyncManager$SyncStatusBroadcastReceiver`, which generates `SyncProgressEvent`s, which are
then interpreted by UI, which increments a progress bar using the incremental value.

In general, event-driven architectures should not require clients to track previous events in order
to infer what to do - all state should be included with the event. This commit makes that the case,
by reporting progress directly as a percentage, instead of reporting increments and expecting
clients to keep track of all previous events.