Genercise client-side incremental sync logic.
ObservationsSyncPhaseRunnable and PatientsSyncPhaseRunnable both had lots of common code, and there
were small inconsistencies between them that went unnoticed. This change centralises the common
logic into a base `IncrementalSyncPhaseRunnable`, which will also be used for Orders in a future
change.

Some notes about the design:

- I'm normally opposed to inheritance, but in this situation, it's syntactically a lot more
convenient, and the expected relationship between the superclass and the subclass is well-defined.
- This is the reason why I've marked the `#sync` method as final and gone for "beforeSyncStarted"
and "afterSyncFinished" callbacks instead of getting subclasses to override the `#sync` method -
it's much closer to a composition-style design in this manner.