Fix bug where Order end time wasn't saving.
We use `stop_millis` to indicate the time when an order should be stopped. `null` has the special
meaning of _'continue the order indefinitely'_. if `stop_milis` was `null`, however, our JSON
request-building code put the value `null` into the specified key, which actually causes the value
to be unset for the request.

Server-side code would automatically set `stop_millis` to `null` if it wasn't set, so this had no
impact on create operations. It does, however, have an impact on update operations, because it's
not possible to change `stop_millis` to `null` once it's already been set previously.

This change uses the special value `JSONObject#NULL` to populate the JSON request instead, which
means that the key is set in the request, which means that the value is set correctly on the server.