Use dark, ancient magic to ensure that seconds are set for form entries.
The timestamp for encounters entered with an Xform didn't include a seconds field previously. This meant that if the same form was filled out twice in one minute, there was no guarantee that the most recently set value would be the one displayed on the patient chart.

This change makes the DateTimeWidget in Xforms use the seconds of the current time by default, and then clears the seconds if the value changes. This workaround gets us a lot closer to the behaviour that the user expects.

Fixes #125.