Fixes #177: "Sorry, form save failed: Index 0 requested..."
ODK keys its form type data off of an `_ID` column, which is a primary key on the `forms` table.
This `_ID`, however, isn't explicitly set anywhere - it's an auto-incrementing primary key.

When a sync occurred whilst a form was being filled out, the form was being replaced in the ODK
database. ODK's FormsProvider, however, models REPLACE operations as DELETEs followed by INSERTs in
a single transaction. This meant that that the `_ID` column changed for every form on every sync,
which means that when ODK goes to look up the form information again on submit, the relevant form
had disappeared.

This change fixes the problem by figuring out the `_ID` of the existing record, and then doing an
`INSERT OR REPLACE` instead of the transactional `DELETE` and `INSERT`.

A better long-term fix might be to make the ODK keep the form information in memory whilst the form
was being filled out, but that's fiddlier.