Add a "Stop order now" button.
Note that I had to make some changes to support when an order gets a duration of zero, in a
circumstance when (e.g.) a user creates an order, administers the order, and then stops the order.
These changes are hacky at the moment because our data model doesn't support < 1 day duration
orders. Long term, we should clean this up, but we have an outstanding bug for that anyway.

Also: Cleanup order fragment XML.

- Remove unnecessary LinearLayout
- Add `android:labelFor` attributes for TextViews that point at EditTexts.