Display count of executed orders in STOP cells when non-zero.
When the user stops an order now, the STOP cell can be a day on which they've already made multiple
executions. We allow for this and display it in the UI, but don't allow the dialog to open, because
we don't want users to mark STOP orders as further executed, and fixing that would push this set of
changes out of the realm of being a hack to something we should fix properly.