Set up unit tests and remove some AsyncTasks.
Previously, all our unit tests were actually instrumentation tests and had to run on real devices.
There were two major reasons for this:
 - Android's support for unit testing has historically been lacking, and so you couldn't set up unit
 tests easily with gradle.
 - The heavy reliance upon AsyncTask throughout the Buendia codebase essentially means we *have* to
 run on real devices, because they require a Main Thread to be available, and a lot of the
 infrastructure around that to be working.

Android now has decent support for unit tests, so we create a unit tests target and run it as part
of our CI (it's pure java, so it doesn't need a device or emulator). Additionally, the codebases'
overuse of AsyncTask has historically been problematic - the codebase employs AsyncTask as heavily
as it uses EventBus, and both of them solve the same problem. We have an awful design pattern where
an AsyncTask is used to do something on a background thread, and then in the onPostExecute, we'd
post the results to the EventBus. Instead, we can just post to the EventBus directly from the
background thread, so we don't need AsyncTasks anymore, which makes our code much more testable.

This change:
 - Moves UserManagerTest over from an instrumentation test to a unit test
 - Changes UserManager to use an executor for background thread computations instead of AsyncTasks
 - Injects an Executor to be used for running code on the background thread app-wide.
 - Runs unit tests on Travis builds.