Display patient sex and age in subtitle on PatientChartActivity
Solution to https://github.com/projectbuendia/client/issues/325

This does leave display of patient and sex and age in smaller font than previously, even if id and name is short. I am uncertain if that is acceptable, given visibility requirements.

A potential alternative would be using a custom view for action bar.